use local pkt_cls.h for new TCA_POLICE attributes
It is a requirement for compiling the PPS code that TCA_POLICE_PKTRATE64 and
TCA_POLICE_PKTBURST64 are present and the build will fail if system headers
are used which don't provide this attribute.

This is achieved by using the local version of pkt_cls.h unless
the system one provides TCA_POLICE_PKTBURST64.

Signed-off-by: Simon Horman <simon.horman@netronome.com>