tc: Add a gact pipe action in tc filter to dump flow stats.
Previously flow stats is dumped using the stats of the first action
in this filter, which will result in problem when 2 filters using a
same police action.

For example, when we apply meter police on both directions of TCP
traffic, the dumped stats is shown same (as shown below):

...,in_port(6),eth(),eth_type(0x0800),ipv4(frag=no), packets:1488557,
bytes:2089059644, used:0.040s, actions:meter(0),9
...,in_port(9),eth(),eth_type(0x0800),ipv4(frag=no), packets:1488557,
bytes:2089059644, used:0.040s, actions:meter(0),6

In order to solve this problem, a gact pipe action could be added at
the first place in each filter to dump the flow stats.

Since gact pipe offload need to be supported in kernel tc (and netdev
drivers if using hw tc datapath), only when the netdev supports gact
pipe offload, this action will be added, otherwise keep the previous
implementation.

Signed-off-by: Tianyu Yuan <tianyu.yuan@corigine.com>